Feedback review verwerkt

+++ AwARe/Assets/FirstParty/Application/Scripts/MainFeatures/RoomScan/Path/LineDrawer.cs
@@ -16,10 +16,10 @@ namespace AwARe.RoomScan.Path
     {
 
         /// <summary>
-        /// draw a line of 'true' values between 2 points on a given grid of booleans.
+        /// Draw a line of 'true' values between 2 points on a given grid of booleans.
         /// </summary>
-        /// <param name="grid">grid of booleans to draw the line on.</param>
-        /// <param name="linepoints"> line to draw. points are coordinates in the grid. </param>
+        /// <param name="grid">Grid of booleans to draw the line on.</param>
+        /// <param name="linepoints"> Line to draw. Points are coordinates in the grid. </param>
         /// <param name="carve">Whether the line should be carved out of a positive grid.</param>
         public static void DrawLine(ref bool[,] grid, ((int, int), (int, int)) linepoints, bool carve = false)
         {
@@ -44,8 +44,8 @@ namespace AwARe.RoomScan.Path
             else
                 c = xdiff / ydiff;
 
-            float b = (float)y1 - a * (float)x1;
-            float d = (float)x1 - c * (float)y1;
+            float b = y1 - a * x1;
+            float d = x1 - c * y1;
 
             for (int x = Math.Min(x1, x2); x <= Math.Max(x1, x2); x++)
             {
